Adjust node security group rules (APIServices) (#62)
Open inbound ports for APIServices: metrics-server, prometheus-adapter
and calico

Remove Control Plane egress rules. The Control Plane SG is added to the
nodes which means ingress is all that matters. We are egressing to a set
of ENIs that have an EKS managed SG and not the one we believe we are
controlling. Everything is working because we allow everything from a
node to the internet.