Add empty main.tf for validation purposes
I think there's value in kms.tf being in standard terraform format,
given its similarity and relation to the base module, so I'm
hesitant to jsonify it and put it in main.tf.json. On the other
hand, having an empty main.tf is kind of silly and I don't love it,
but it also doesn't hurt anything. :-/